Remove TDMLogger
Previously, tala.log.TDMLogger was needed in the 'tala' repository
since the logger was injected into some of the build components, to be
used by other components downstream.

However, no logging was done in downstream verification or generation
components, so TDMLogger was actually not needed in the 'tala'
repository.

This change removes tala.log.TDMLogger, and all injections thereof,
from the 'tala' repository, migrating it to the 'tdm' repository.

Taiga story #1202

Change-Id: I299ae08d75cf6ca8877a0cc5f1afdc89c97ccfc2