Migrate tdm.lib.tests to tala.model.tests
Previously, tests for the model classes were located in 'tdm' because of
improper unit testing and dependencies toward code in 'tdm', e.g. the
DDDComponentManager.

This change ensures that dependencies toward code in 'tdm' is removed
and that the tests are moved to 'tala.lib.tests' instead.

Additionally, this change moves 'tala.lib.move_factory' to 'tdm' since
it is only called by code in 'tdm'. Similarly, the 'MoveFactoryTests'
class is extracted from 'tala.model.test.test_move' and moved as well.

Taiga story #1202

Change-Id: Ib0f24ef7f1654dbc663d99f4d792068f12d97c38