jenkinsfile-change-merged: Archive properly
Previously, the Jenkins job that ran on merged changes did not archive
properly because a 'tala' wheel was never built.

It's tempting to build and use a 'tala' wheel during unit and
integration tests. However, one runs into problems because
'tala/installed_version.py' can not be found in the repository. This
happens because 'pip install' only ensures that 'installed_version.py'
is created in its install-location. When pytest is run on the
repository however, the test discovery mechanism forces the local
repository onto pytest's 'sys.path', letting the local sources take
precedence over the installed version.

As a strategic decision, it makes sense to install in editable mode
when running unit and integration tests, since they test internals and
the test code sits next to the sources, but are not included in the
built wheel.

This change ensures that a wheel is built, so that there is something
to archive. It doesn't install the built wheel however, to avoid the
problems mentioned above.

Taiga story #1202

Change-Id: I4d88ead12870f3bec7971a592d3f89e2eaf9844c