Subscribe to submodule 'internal_ddds'
This change adds 'branch = .' for the 'internal_ddds' repository in
the .gitmodules file as an instruction for Gerrit to subscribe to
its changes.

'branch = .' means that branches with the same name are
subscribed. For instance, if the 'master' branch of 'internal_ddds'
is updated, Gerrit will automatically update the 'master' branch of
this repository, if allowed to do so.

For more details, see:
https://gerrit-review.googlesource.com/Documentation/user-submodules.html

Change-Id: Iea686de00052b5c46f97e0910a6b225f6d2d41b6