jenkinsfiles: Add pylint checks
Before this change flake8 was used to run static code checks, but the
number of checks run by its error checker 'pyflakes' was few, see
http://flake8.pycqa.org/en/latest/user/error-codes.html.

On the upside, flake8 ran very quickly.

This change adds pylint to the test routine, providing a framework for
many more checks, although at the cost of speed. For all checks, see
'messages' sections at https://docs.pylint.org/en/1.6.0/features.html.

In this initial version, only 3 out of all features are checked, taking
about 15 seconds of time and succeeding on the current state of the
repository.

The 3 checked features are:
- import-self
- reimported
- wildcard-import

Change-Id: I8543ff1f0494a1c22d60798cef0ac901aeddb160