Clean up JSONLoggable implementation
When the 'JSONLoggable' class was introduced recently, two design
patterns were used to achieve JSON-compatible log entries. On one hand,
objects were checked for the 'can_convert_to_json' attribute, which in
case it existed was queried before attempting to convert to JSON. On
the other hand, the JSONLoggable class ensured in itself that an object
could be converted.

This change removes the 'can_convert_to_json' attribute in favor for
the JSONLoggable class.

Taiga story #1320

Change-Id: I2a960ab7fb45e661dc9c23ffccffb04375491ad7