Dockerfile: Remove unused escape characters
Before the Dockerfile was added, it was pasted inline in a Jenkins job.
At that point it needed to escape backslash characters.

When it was added as 'Dockerfile', the escape characters remained,
causing errors because they ended up in the actual terminal command.

This change removes them, solving the problem.

Change-Id: I1c376a19c63b409db01ec38cf41c39733d6f1471