jenkinsfile-release: Properly activate twine venv
Before this change, a virtualenv was created to install twine, needed
for uploading to PyPI. The virtualenv was however not activated
properly before the upload took place, likely causing the markdown
README to not format properly on https://pypi.org/project/tala.

This change ensures that the virtualenv is activated properly so that
twine has access to the markdown formatter when uploading, fixing the
problem.

Change-Id: I12dc978fb5b7554bc7c117b0c2d84d47e732174e