jenkinsfile-release-latest-docker: Checkout tala
Before this change, Jenkins pipeline
'jenkinsfile-release-latest-docker' was defined inline in its Jenkins
job, and it used an inline Dockerfile. When the pipeline was migrated
from its job to 'jenkinsfile-release-latest-docker' recently, the
Dockerfile was also extracted to its own file, 'Dockerfile' inside this
repository.

Additionally, before this change, 'jenkinsfile-release-latest-docker'
copied the tala wheel from the 'tala-latest' Jenkins job, which meant
it never needed to checkout the repository.

Since 'Dockerfile' recently became part of the repository however but
no checkout was done, the job would fail because the Dockerfile was not
available.

This change does a checkout of the repository and builds the tala wheel
itself, instead of copying it from 'tala-latest'. This allows it to
access the Dockerfile, fixing the problem.

Change-Id: I1ada1edaa51c50ec9ce18b04300631cedfb55739