Add 'tala interact'
Before this change, tala users could not use tala to interact with and
test the behaviour of their DDD when it was deployed to a TDM server.

Several tools existed, but they all had external dependencies, such as
Android apps and test tools for HTTP endpoints.

This change adds command 'tala interact', which allows them to interact
with an as-simple-as-possible command line interface.

The new command:
- handles signals SIGINT and SIGTERM when the command needs to be
stopped by force.
- uses the interaction testing format; stdout history can be pasted
straight into an interaction test.
- stores stdin history in './.tala/history'; pressing the up key scrolls
through history.
- supports passivity; the user is considered passive when nothing has
yet been typed.

Additionally, this change:
- adds 'tala.utils.observable', which defines inheritable classes
'Observable' and 'Observer' for the observer pattern.
- adds 'tala.utils.tdm_client', a basic HTTP client for TDM.

Limitations:
- With the new command, it's always the system that takes the first
turn.

Taiga story #1545

Change-Id: Ia14d9166ce01caddeadfe30a25e032353b48182d