Create deployments config for 'tala interact'
Before this change, 'tala interact' required a mandatory 'url'
positional argument to be provided with the command every time, e.g.
'tala interact https://my-ddd.mycloud.com:9090/interact'.

This change adds a new deployments config, by default
'deployments.config.json', where 'tala interact' looks for an
environment with the url instead.

After this change, the full command is as simple as
'tala interact environment', where environment is one of the keys in the
deployments config. When a deployments config is created, there's a
'dev' environment in it, allowing developers to run 'tala interact dev'.

Taiga story #1545

Change-Id: I278e32d5e54fd68651f172b60247336ac9aa4e6f