RASAGenerator: Use markdown
Before this change, the RASAGenerator class generated RASA training data
in the JSON format. The JSON format is however syntax heavy and hard to
read and inspect for humans.

This change converts the RASAGenerator class so that it generates in the
markdown format instead, which is easier to read for humans.

Change-Id: Ie967a13d08f031e8b9d2028f91f0b3a015339a88