Add new built-in sort 'person_name'
This change adds a new built-in sort 'person_name'. Support for
generating Rasa training data with answers for this sort is also
added.

NLG for the new sort is only supported for non-RGL grammars.

To improve extensibility for built-in sorts, the following
additional changes are made:
- When it comes to system answers that use placeholders,
  tala.nl.gf.auto_generator now generates GF functions for all
  built-in sorts, rather than for an explicit set of sorts. This
  means that auto_generator should no longer need to be changed
  when introducing new built-in sorts that use placeholders for
  NLG.
- A new method Sort.value_as_basic_type is added, returning the
  passed value as a built-in Python type. E.g.,
  PersonNameSort.value_as_basic_type(PersonName("Anna")) returns
  "Anna".
- A new method Sort.value_from_basic_type is added, returning
  the passed instance of a built-in Python type as a proper value
  for the sort. E.g., PersonNameSort.value_from_basic_type("Anna")
  returns PersonName("Anna").
- Sort.value_as_json_object now uses Sort.value_as_basic_type by
  default.

Furthermore, tala.ddd.ddd_py_compiler now creates instances of
built-in sorts using BuiltinSortRepository.get_sort(name) rather
than BuiltinSort(name); this ensures that polymorphism for built-in
sorts works as intended.

Taiga story #1553

Change-Id: I1fa6397d79e8d8c3eb9ce177ea089a13b0acd9d7