Rasa NLU: Use trainable entity identifier
Previously, 'tala generate-rasa' generated training data with entities
on the forms 'sort:value' and 'predicate:value'. This worked well when
the training data was in the JSON format.

Since the recent migration to the markdown format however, when parsing
with a Rasa NLU trained on these forms, it for some reason returns
entities where the identified entity is 'sort' instead of 'sort:value'.

This change replaces the form 'sort:value' with 'sort.value' instead,
avoiding the problem. 'predicate:value' is similarly replaced by
'predicate.value'.

Change-Id: Iedea934fce156945663e4e649a8455685f40b572