Rasa NLU: Don't treat builtin sorts as entities
Before this change, all sortal slots in the grammar would be generated
as entities in the Rasa training format.

Since builtin sorts should be recognized by pre-trained NERs instead of
being trained on, it doesn't make sense to generate them as entities.

This change ensures that they are not generated as entities.
They are still generated in intent examples however, which is needed to
achieve good intent performance. This means that DDD developers should
still include them as sortal slots in the grammar.

Change-Id: I6d19d11b9e46377e9422c3b49dbc8341d1d5ba80