jenkinsfile-change-merged: Push Docker image
Before this changed, there was one 'jenkinsfile-change-merged' which
ran the full test suite whenever a change was merged to master.

Simultaneously, there was another 'jenkinsfile-release-latest-docker'
which built and pushed the 'nightly' Docker image every night.

This change migrates the Docker release to 'jenkinsfile-change-merged'
and removes 'jenkinsfile-release-latest-docker' altogether. From now on,
a 'latest' release will be made to Docker as soon as the full test suite
passes after a change is merged to master.

Note that this change renames the 'nightly' tag to 'latest'.

Change-Id: I1581d6823a38d72c920a0bc59549a762d7be4383