jenkinsfile-change-merged: Push candidate
Before this change, the jenkinsfile-nightly pushed a 'latest' docker
image after all tests had succeeded. However, the 'latest' tag should
be considered 'the latest release'.

After this change, a 'master' image is pushed if a wheel can be built on
the master branch, and a 'candidate' image is pushed if all tests
succeed.

The 'latest' tag is no longer pushed.

Change-Id: Ifcd64f8c142e4e480d237ef89347feec64470807