TDMClient: Remove session management
Before this change, the TDMClient handled sessions, but was limited to
a single session. This limitation was fatal for the Twilio integration
service, which needed to handle several concurrent sessions.

This change extracts session management from the TDMClient, to the
caller, allowing the Twilio integration service to handle several
sessions.

Whenever the TDMClient is called, the session needs to be injected into
the client, and will be forwarded as-is to TDM.

Change-Id: I3cb6d42df7cc455dd813706bc5b2f888e1cfbb3c