jenkinsfile-release: Push existing image
Before this change, jenkinsfile-release built a new image.

This change ensures that the existing image, built by
jenkinsfile-change-merged, is simply re-tagged with the version instead.

Additionally, the PyPI username and password are set as environment
variables instead of passing them in the sh command. This potentially
solves an authorization problem encountered with Twine, where access is
inexplicably denied.

Change-Id: I6399641ae82a383ca469f3db75850c0ad824c499