jenkinsfiles: Properly fetch Git hash
In a recent change, the Git hash to be tagged on the Docker image was
retrieved using the 'GIT_COMMIT' property of the returned value of the
'checkout' step.

For some reason, when the 'checkout' step is checking out a tag, the
'GIT_COMMIT' property points to the commit hash of 'master' instead of
the hash of the checked out tag.

This change ensures that the commit hash is fetched by an explicit call
to 'git rev-parse HEAD' instead, which retrieves the correct commit
hash.

Change-Id: Ica34a893d3fb05923b001d9c2ddd727cdca6e187