interact: Accept deployment URL
Before this change, 'tala interact' required a deployment URL to be
added to a deployments config in order to be used.

This change allows the URL to be provided directly in the
'tala interact' command. Previous behaviour is still supported.

Additionally, this change updates 'tala.config' so that:
- Strings are formatted new-style (with "{}".format(x))
- Paths are wrapped in 'pathlib.Path' objects

And 'tala.test.test_config' is updated to work with pytest, dropping
Python's 'unittest' support.

Change-Id: I1c21595ed8cb6f2321a812df8ec2ad904c678c81