Change 'tala generate-rasa' to 'tala generate'
Before this change, 'tala generate-rasa' was a stand-alone command that
generated training data for Rasa NLU. This was fine, since Rasa was the
only supported NLU. Soon however, Alexa and its NLU needs to be
supported too.

This change turns the 'generate-rasa' command into the more generic
'tala generate' and lets the platform become an argument instead. This
enables the same command to generate training data for both platforms.

Taiga story #1630
Taiga task #1679

Change-Id: I13417af081c0d03ad4ab3cc990fa706eca4f7ed8