Add 'tala generate alexa'
This change adds 'alexa' as a format to the 'tala generate' command,
generating NLU training data on Alexa's format, and compatible with the
TDM runtime.

In case several DDDs are used, the DDD name is encoded into intent and
type names so that TDM knows where to look for them. All DDDs however
need to be served by the same TDM deployment.

Taiga story #1630
Taiga story #1148

Change-Id: I9c9c00fdaa8c960af505f8bce008db56101d59b0