Support cyclic surface variation for reports
This change adds support for cyclic surface variation for
reports in non-RGL DDDs by
- supporting multiple report variants for the same set of bound
  parameters
- supporting a new optional attribute "selection" for <one-of>
  elements in <report>, with supported values "disabled"
  (default) or "cyclic"

Support for honouring variant order is achieved by creating
a GF file with "probabilities" with values refelecting the order
defined in the DDD grammar. Note that this file is now always
created, also for DDDs with RGL enabled.

Additional changes:
- Unit tests for GF code generators are extended to cover file
  creation.
- Unit tests for tala.utils.as_json are added.

Taiga story #1594

Change-Id: I2b3fd21fc7a26e991bfc048e50a7b69bd9cadeab