jenkinsfile-quick: virtualenv -> Dockerfile-dev
Before this change, a virtualenv was used in jenkinsfile-quick to
install dependencies and run unit tests. A recent update to pip however,
revealed a Jenkins confusion where the virtualenv-local pip is executed,
but runtime files are imported from the system-global pip.

This caused a fatal exception to all pip commands when the virtualenv
used version >= 19.3 and the system-global pip had version < 19.3.

This change removes virtualenv from jenkinsfile-quick altogether,
replacing it with Docker container agent, built from a new Docker file:
'Dockerfile-dev'.

Change-Id: Ia070b770f5942eea5674553a6361e5450c020cb0