jenkinsfiles: virtualenv -> Dockerfile-dev
Before this change, a virtualenv was used in jenkinsfile-change-merged
and jenkinsfile-release to install, package and test tala and its
dependencies. A recent update to pip however revealed a Jenkins
confusion where the virtualenv-local pip is executed, but runtime files
are imported from the system-global pip.

This caused a fatal exception to all pip commands when the virtualenv
used pip >= 19.3 and the system-global pip had version < 19.3.

This change removes virtualenv from jenkinsfile-change-merged and
jenkinsfile-release altogether, replacing it with Docker container
agents, built from 'Dockerfile-dev', in the stages that need it.

A new 'Dockerfile-release' is created with the Python dependencies
needed to release Tala to PyPI with jenkinsfile-release.

Additionally, jenkinsfile-change-merged now archives the generated
'requirements.txt' as an artifact.

Change-Id: I32f4515c8020dca85ef3cd30661c086847702b66