jenkinsfiles: Use one agent at a time
Before this change, the jenkinsfiles declared an agent on the pipeline
level, then another agent in some of its stages. This caused two
different agents to be allocated during some stages of the build.

This change sets 'agent none' on the pipeline level and sets the agent
explicitly in its stages instead, effectively preventing more than one
agent to be allocated simultaneously.

This possibly prevents an intermittent failure when two different
executors of the same agent were running the same build concurrently.

Additionally, this change replaces the plugin-based 'WsCleanup' step
with the builtin 'deleteDir()' step.

Change-Id: I395925f8a18c3610adf931487d85b8af82502f1c