jenkinsfiles: Properly copy files between stages
Before this change, in all jenkinsfiles, different stages were being run
on different agents. An assumption was made that the filesystem at the
end of a stage would be maintained to the beginning of the subsequent
stage, even if the two stages took place on different agents. This
assumption was however false, for instance causing multiple wheels to be
archived at the end of a build-wheel stage, although only one wheel had
been built.

This change ensures that important and shared files are stashed at the
end of a stage. It also ensures that each stage clears its working
directory before unstashing shared files, to avoid conflicts.

Additionally, this change replaces the builtin 'deleteDir()' step with
a 'sh "rm -rf"' step throughout all jenkinsfiles, since 'deleteDir()'
removes the entire workspace on docker agents, preventing them from
running.

Change-Id: I861a8e83eabdc8b7958178b7e1892eb9f3bb1040