jenkinsfiles: Reuse agents
Before this change, stages of all jenkinsfiles were considered to be run
on different agents. This fact required the filesystem to be cleaned and
setup at the beginning of each stage, which added additional steps and
was in itself fragile, since future changes could easily forget to
comply.

This change takes advantage of the 'reuseNode true' instruction on
stages that use 'docker' and 'dockerfile' agents. This ensures that all
stages run on the same filesystem. In case of 'docker' and 'dockerfile'
agents, they share the filesystem with the host agent; in other stages
the host agent runs them itself. As a result, the clean and setup steps
in the beginning of each stage are removed, along with the usage of
'stash' and 'unstash' steps.

Change-Id: I5c1b7a71b096931d3743ccdd90e5afafa225098c