Move: Add is_realized()
Before this change, the Move class had an internal check whether or not
the move was realized or not. It was however impossible for external
callers to do the same check.

This change adds an is_realized() method to the Move class, so that both
external and internal callers can do the check.

Change-Id: Idfeca9f03869b2d6255a30ff4440e3a3ff1111ea