AutoGenerator: Allow answers with just slot
Before this change, when a non-RGL grammar had a system answer with a
single slot but no text prepending or appending the slot, e.g:

 <answer speaker="system">
   <slot type="individual" predicate="dest_city"/>
 </answer>

the natural language GF file would be generated with a syntax error.

This change ensures that such system answers are allowed, avoiding the
syntax error.

Additionally, this change verifies that such system answers are already
supported by RGL.

Note that YAPF automatically formatted the changed files, which obscures
the actual change.

Change-Id: Ic18746bfb9de8f40cb387061aa5f03a0e597a426