Move: Properly implement __eq__
Before this change, the __eq__ method in the Move class had some
shortcomings. For instance:
- objects of subclasses could be considered equal with objects of other
subclasses, although they weren't identical.
- If a subclass did not override the __eq__ method, and had additional
attributes compared to Move, then those attributes would not be checked.

This change ensures that __eq__ is implemented according to current
python recommendations, checking both the class and all attributes, as
found in the object's __dict__. The implementation is done in a new
mixin, 'tala.utils.equality.EqualityMixin', for reusability.

Change-Id: I5101a9906573f70eaa693fcb3c0084af5c936c58