Remove incorrect changelog entry
A recent change added a changelog entry saying that 'tala verify' fails
with a syntax error, supposedly when the generated GF files are
compiled.

This is incorrect since Tala does not use GF. 'tala verify' indeed
generates the syntax error, but doesn't compile with GF so no error will
be discovered.

This change removes the changelog entry.

Change-Id: If7afc00609cd96d432de37885667ad16d976a795