tala generate: Fail fast for unsupported language
Before this change, when running 'tala generate' on a language that was
not supported in the backend config, a cryptic error message was given,
with no clues on how the problem could be fixed.

This change ensures a better error message is provided instead.

Additionally, this change modifies the BackendDependencies classes, so
that DDDs are loaded by the concrete classes instead of the
AbstractBackendDependencies class. This is needed by TDM to provide
similar error messages.

Resolves issue #1642

Change-Id: I826a012e7dcf1e0b0928a1ce463ba30f88493c23