Make most config fields optional
Before this change, fields in config files were always required
load-time; if a field was missing when a file was read, the
field was automatically added to the file (with a default value)
and an exception was raised.

After this change, only some fields are required load-time (see
below), while most fields are optional with implicit default
values; if a field is missing when a file is read, its default
value is used, and no exception is raised.

One of the motivations behind the change is that previously,
when a new supported field was added to a config class,
the extension of the config format was backwards-incompatible in
the sense that an action was required from developers for
migrating their config files to the new format. Another advantage
of optionality is that it can hide configuration details that are
irrelevant or uninteresting for the developer.

Required fields after the change:
- for backend configs:
 * ddds
 * active_ddd
 * supported_languages
- for DDD configs:
 * use_rgl

Additionally, tala.utils.text_formatting is removed since its
only method wasn't very useful (it didn't support noun-verb
agreement) and since it's no longer used.

Change-Id: Id01273821bb932c4648a65ee7f3122d6b886ca86