Ontology: Prevent ambiguous sorts and predicates
Before this change, when a sort and a predicate of the same DDD had the
same name, 'tdm build' failed upstream with a mysterious and seemingly
unrelated GF syntax error.

This change ensures that en exception is raised when a sort and a
predicate share the same name in the same ontology, failing fast and
properly informing the user.

Part of issue #1544

Change-Id: Ie8333f64e8cdf87184b4d4587fbcbbe5c938234f