tala create-ddd: Prevent illegal DDD names
Before this change, 'tdm create-ddd' and 'tala create-ddd' had different
implementations.

'tdm create-ddd' prevented illegal DDD names, but 'tala create-ddd' did
not, causing errors with GF downstream if for instance a dash was part
of the name.

This change ensures that the implementations no longer differ, and that
both of them prevent illegal characters.

Resolves issue #1258 (again)

Change-Id: Icfe333d6d7fc2d1f41230b0e09d6bdf1b1e505dc