Move: Generate proper semantic expressions
Before this change, the Move class was an AsSemanticExpressionMixin
class but the semantic expressions that the different Move subclasses
generated were long and hard to digest.

This change ensures that they generate short and concise semantic
expressions instead.

Before the change, a realized request move could be generated into this:

  Move(answer(paris), ddd_name='mockup_ddd', speaker=SYS,
  understanding_confidence=1.0, weighted_understanding_confidence=1.0,
  perception_confidence=1.0, modality=speech)

whereas after the change, it would be generated to:

  answer(paris)

Change-Id: I0800907980c75deee0ffe01bdce52cb236ac38ef