Relax XML parser for plans
Previously, the DOM parser of the ddd_xml_compiler.py took responsibility
for checking that some elements of the plan XML didn't contain other
elements than the expected ones. This responsibility has earlier been
transferred from the parser to the schema validator. In addition to
assuming this unnecessary responsibility, it also refused to accept
whitespace (spaces, newlines) as child nodes. The DOM parser now accepts
whitespace, and doesn't try to duplicate the validation task for
"condition" and "forget" element. It is also tested that the schema
validation actually happens.

Change-Id: I8bc79da4e899c7a67c3eff7e1ba214023797b7cf