Introduce the assume_system_belief plan item
This change introduces the assume_system_belief plan item (previously
known as the assume plan item). The result of assuming a proposition as
system belief is that it is added to the system's private beliefs.

Also adding tests for validation of assume_shared plan item and minor
refactoring of method names in ddd_xml_compiler.

Change-Id: I254205a89009b0247d4973e9e3b956b9c91162ad