ICMMove: Add is_grounding_proposition()
This change adds a method 'is_grounding_proposition' to ICMMove objects.
It returns True when an ICM move is grounding a proposition, i.e:
for icm:und*pos and icm:und*int with content, and False otherwise.

This change is needed upstream to determine which propositions in
private.bel / shared.com that are currently being grounded.

Change-Id: If84c7aecb933f3b7ec90d63c7cc56d244a4772d3