tala generate rasa: Add DDD name for entities
This change adds the DDD name when generating the training samples
containing custom sortal and propositional entities.
This is needed so that TDM can know which DDD the entities that are
returned from Rasa NLU belong to.

Additionally, this change fixes a problem where integration tests
for 'tala generate rasa' succeeded for the wrong reason.

Taiga task #1737
Taiga story #1677

Change-Id: I6a6dcb2c2a44fce5459ce2d08db9a4b778e8803e