GoalProp/PredProp: Add method as_move
This change adds a method 'as_move' to the classes 'GoalProposition' and
'PredicateProposition'. It is needed upstream in TDM to convert such
propositions in TIS.NextUtterance.alts to move expressions such as
'request(call)', 'ask(?X.selected_contact(X))' or
'answer(selected_contact(contact_john))' for the 'expected_input' object
in TDM responses.

Change-Id: If5923f75fc182ea90432f67c2ff2ed012e9dc7c2