tala test: Improve semantic interactions
Before this change, semantic input in interaction tests could only be
specified as semantic expressions. Explicitly specifying both the
perception and understanding confidence scores could not easily be done.

This change enables rich semantic moves as user input, allowing both
perception and understanding confidence to be specified:

  U> [{"semantic_expression": "request(call)",
     "understanding_confidence": 0.55, "perception_confidence": 0.79}]

Furthermore it enables interpretations to be specified too, allowing an
utterance and the modality to be specified:

  U> {"moves": ["request(call)", "answer(contact_john)"], "utterance":
     "an utterance", "modality": "speech"}

or

  U> {"moves": [{"semantic_expression": "request(call)",
     "understanding_confidence": 0.55, "perception_confidence": 0.79},
     {"semantic_expression": "answer(contact_john)",
     "understanding_confidence": 0.55, "perception_confidence": 0.79}],
     "utterance": "an utterance", "modality": "speech"}

Change-Id: Ica9de892379dfc6fee9d083911ff84679d074d9b