Create 'tala endurancetest'
Since 'tala test' was created and 'tdm test' removed, 'tdm
endurancetest' could not succeed as it was using the files for
interaction testing in Tala and those need to work with a served
TDM pipeline.

This change adds 'tala endurancetest', which allows endurance
testing using a remote TDM pipeline so DM and custom NLG
components can be tested altogether.

This change resolves #1769. Related to #1768.

Change-Id: I949d35017f0245c18ab50c8303164da362274e01