tala test: Wait for TDM
Before this change, when 'tala test' could not connect to TDM it
failed fast with a fatal exception, causing Jenkins builds to fail
intermittently because of a race condition between 'tdm serve' and
'tala test'.

This change ensures that 'tala test' waits up to 3 seconds,
attempting to connect every 0.1 seconds, preventing the
intermittent Jenkins failures.

Additional changes:
- 'tala test' prints a line at the beginning of each test session:
"Connecting to http://localhost:9090/health..." which ends with "done"
when the connection is established, or "failed" if it fails and times
out.
- 'tala test' now properly prints the "Running interaction tests from
path/to/interaction_tests_eng.txt" line appropriately when tests from
the given file are about to start, instead of in the beginning of the
test session. This means that running tests from more than one file at
once, prints as expected:

  Running interaction tests from path/to/interaction_tests.txt
  ....
  ----------------------------------------------------------------------
  Ran 4 tests in 0.249s

  OK
  Running interaction tests from path/to/interaction_tests.txt
  ....
  ----------------------------------------------------------------------
  Ran 4 tests in 0.232s

  OK

Change-Id: I0dbf9fdb149a4b2964d73e8194057b519d4f723c