Refactor http_formatting
Since http_formatting has no more dependencies to the DDD or any other
major components and does not longer have state, it is no use having it
as a class.

This change turns the methods of the HTTPFormatting class into
functions in the http_formatting module, and removes the class.

Change-Id: I8657ceda782418486d8d24bf262d04d61fc786ef