Fix: Commitments serialization/deserialization
We have spotted errors where a TypeError has been raised caused by the
following line in commitments.py (line 125 in this changeset):

if turn_number + self._previous_commitment_lifespan < current_turn_number:

The exception is "TypeError: can only concatenate str (not "int") to str"

We have not been able to reproduce the error in tests, but have made
some changes that we believe will make sure this does not happen again.

The changes are first and foremost that we are catching any exception
raised in the offending method. We are also making sure to interpret all
the values as integers using 'int()'.

Catching all exceptions means that we're effectively swallowing all
error output when this happens. This should be further investigated and
fixed (indicated by comment in code).

Change-Id: I552c4c4dd9959d26ab1870c9baf23a3b0e2af6f3