Treat DDD specific components as DDD
Previously the bundle of a DDD (domain, ontology and service interface),
a parser and a parameter retriever, was referred to as "DDD specific
components", especially in the DDDComponentManager class. This is a
clumsy name for something which is essentially treated as the actual
DDD.

This change updates the names of the access methods in DDD component
manager to refer to DDDs instead of DDD specific components (e.g. the
method "get_components_for_all_ddds(self)" is now named
"get_all_ddds(self)", and the method
"get_ddd_specific_components(self, name)" is now called
"get_ddd(self, name)".

Change-Id: Ie9235c9cfed7fedf61e0ec4cc5c737353944a37b