Refactor Class Names
This change alters two class names in order to make their roles clearer.

The class DDDSpecificComponents, which consists of a DDD along with a
parser and a parameter retriever (two classes which are created from DDD
components) is renamed into ExtendedDDD, which clearer reflects what it
is.

The DDDComponentManager, which consists of a collection of ExtendedDDDs,
is renamed into DDDManager, which better reflects its use and function.

The ComponentLoader is accordingly renamed ExtendedDDDLoader, and the
ComponentSetLoader renamed ExtendedDDDSetLoader.

This is a breaking change, and will be a major version upgrade.

Change-Id: I36991dd084a3bbc286096f3bec6a14fb9c1699f6