Task 925: Track num calls last hour in HandlerUserRates
Previously, the call rate handler in tala.utils.user_rates was not
tracking the number of calls for a time unit. This change adds this
kind of tracking, so that the database is keeping track of the number
of calls to a certain offer in the last n minutes.

Also:

  - Add unit tests for user_rates

Change-Id: I327995c2d266c6d36673d2ab932be1548ac94943