Jsonify fields with invalid types for Azure table
Before this change, when creating/updating an entity with
"CallsLastPeriod", the update to the table DB would fail because array
is not a valid data type for an Azure table.

This change adds some methods to jsonify data in "CallsLastPeriod"
before updating it to the table. Then when querying this data it is
made into a list again.

Change-Id: I90101223a2a45e72d407034e49c231ba019f9c31