Support timing constraints in interaction tester
Previously, the interaction tests only cared about output, and never
about the speed of the interactions. There is a need to check if
interactions are slow, and when they become slow.

This change adds the possibility of adding time constraints to
interaction tests. The constraints are on three different levels.

 1. Test: Define a maximum running time for the interaction.
 2. Average time per user turn.
 3. Maximum length of the longest turn of the interaction.

All constraints are optional, and are added to the interaction tests
JSON file.

Change-Id: Ie8034810b756f061a355c35ae911f3a4504715fd