Change name to tala.testing.interaction
When the interaction test framework that we use today was introduced,
there was already an interaction test framework called
"tala.testing.interaction_test", which motivated the "_noveau" suffix of
the new framework. In a recent change, the old framework was removed.

To make the name less clumsy and to comply with python package naming
conventions, the package is renamed from
tala.testing.interaction_test_noveau into tala.testing.interaction

Change-Id: I93726c3d6697f252b3d04225b9571fdcd850dc22