Allow integer strings as individual names
Previously, integersas individuals were earlier
unconditionally interpreted as integers. This is not always
practical, as some APIs name entities using integers. To facilitate
integration of such APIs, we want to allow  individuals to be named
using integer strings.

This change allows parsing propositions like "destination(35362)", and
moves like "answer(destination(3739))".

Short answers using these strings ("answer(6532)") are however not
allowed. Such a move is unconditionally interpreted as containing an
integer individual.

The change is also valid for floats. It is not tested for other
builtin sorts.

Change-Id: I63a199761e5e5bd3acf2729b70dcb1171fedb24d