Catch "ZeroDivisionError" in interaction tester
Before this change, we would occasionally get a "ZeroDivisionError"
when attempting to calculate the average turn times of an interaction
test because the onset_times list would be empty. This error could
probably be masking other errors, potentially related with not
receiving streamed data.

This change captures this error type when it is thrown and attempts to
show some potentially useful data in a warning.

Change-Id: I5968575e537d2e4305b01839d7926eb159646dd0