Add timeout to MQTT Client
Previously we have had errors in streaming - probably in flushing the
streams, causing the streaming sessions not to be closed. In order to
gain control over this, and understand when this happens, we add a
timeout.

If the streamer thread doesn't terminate after 3.0 seconds after
flushing it, we continue closing the streaming session anyway, and log
a warning about it.

Change-Id: Ibd2a5047260c3a27ce15f85f9f41dc2a7e0e2d28