Ensure items are streamed
In order to make it easier to identify where streaming errors happen,
this change makes the MQTT client listen to the MQTT event stream,
raising an exception if a message is lost.

This is the first step, and should be followed by some kind of
mitigation if the problem arises.

Change-Id: Icb67bdce24c20784c5c3f9d593f403ec1f0906af