MQTT: Compare JSON
Previously the strings "är" and "\u00e4r" were considered different
by the MQTT streamer, causing an error with verification.

A previous change added a normalized comparison function for that
purpose, which turned out not to work in practice.

This change loads the JSON strings where the abovementioned strings were
included and does the comparison on the resulting dicts instead, which
eliminates the encoding issues.

Change-Id: I91fdb4852e97d3616c7e59b81f31b67fe94af075