Ensure that tala.model.openqueue.Interpretation works
WHEN SUBMITTING THIS CHANGE WE NEED TO CREATE A MINOR RELEASE.

Previously the methods first(elem) and last(elem) of Interpretation were
non functional. This change makes sure that the methods work and that
they are tested and also adds basic tests for Interpretation.

It also adds properties for first_element and last_element for both
Interpretation and OpenQueue, and deprecates the methods for first() and
last(). (The names "first" and "last" would have been better to use for
the properties, but since they are in use with a different meaning,
we do it this way.)

Change-Id: I60cd90be8b05affd575c443f5f008006e4f6a98e