Add "once" plan construct
This change adds the "once" tag to the plan XML language. Plan items
inside the "once" element will execute only once.

It does so using a special purpose boolean proposition, which is set
when the content of the element has been run. Once the proposition has
been set, the content is not run again until a "forget_all" has
executed.

Change-Id: I741f89b1d62f66046569936fa5702f45be101085