Change Once
Previously the "once" plan item made the assumption that it was executed
after the execution of the entire content. As plan execution is not
entirely linear, this may not be what we want.

This change makes the assumption be made as soon as the once plan item
had been entered.

Change-Id: Ib16a308d3a7a79f00066b2fb178753a075150858