Integrate exceptions i InteractionTest output
Previously when an exception was raised during interaction test
execution, execution was stopped, and an error response was generated.
This lead to confusion and very little information to the developer
about what went wrong in many cases.

This change treats raised exceptions as something natural in testing,
and generates output for the test execution as for other tests.

Change-Id: I0226da87a3f5249eb2713ec20dd9da3bb81e0d6b